requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
phonenumbers==8.13.40
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
